// frontend/src/screens/special/ui/AnyQuestions/ui/Form/lib/types.ts
export type TFormData = {
    fieldName: string
    fieldPhone: string
    fieldEmail: string
    fieldComment: string
    checkboxConsentsTerms: string
    checkboxSubscription: string
    buttonSubmit: string
    successTitle: string
    successDescription: string
}

export type TFormFields = {
    fullName?: string
    phone?: string
    email?: string
    comment?: string
}